Immediately add a pending status on PR
Instead of waiting until we get to the PR request to set status we will
immediately set status on it to pending so users know the service is
up and running.  The status message also puts the # of tests that are
ahead of this one, eg. "CI requested, #waiting = %d"

Signed-off-by: Tony Asleson <tasleson@redhat.com>