testlib: Check client certificate
We are doing 2 way certificate verification.  However, we also
need to check the certificate itself as our built-in CA will
trust client provided certificates that have been appropriately
signed by a trusted CA.  For this to be secure we need to
be confident that no one can create a certificate we trust with
the same serial number.

Signed-off-by: Tony Asleson <tasleson@redhat.com>